pyproject.toml
[project]
name = "gauge-3d"
version = "0.1.0"
description = "3D reconstruction and distance measurement from video"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "tqdm>=4.65.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "diffusers>=0.20.0",
    "accelerate>=0.20.0",
    "safetensors>=0.3.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "scikit-image>=0.21.0",
    "einops>=0.6.0",
    "timm>=0.9.0",
    "requests>=2.32.4",
    "pyyaml>=6.0.2",
    "huggingface-hub>=0.33.1",
    "python-dotenv>=1.1.1",
    # 4D Gaussian Splatting dependencies
    "open3d>=0.18.0",
    "plyfile>=0.7.4",
    "roma>=1.2.0",
    "kornia>=0.6.12",
    "trimesh>=3.21.0",
    "psutil>=5.9.0",
]
